Can draw bounding boxes with text

+++ Cooop365ML/Models/RoboFlowResponse.cs
@@ -11,8 +11,8 @@ namespace Cooop365ML.Models
     {
         public double x { get; set; }
         public double y { get; set; }
-        public int width { get; set; }
-        public int height { get; set; }
+        public double width { get; set; }
+        public double height { get; set; }
         public double confidence { get; set; }
         public string @class { get; set; }
     }
